Try harder to generate deps for Fortran module files
The Fortran Scanner finds INCLUDE and USE statements, and treats them the
same - no dependency is generated if the indicated file does not exist.
This is correct for include files, but not for module files - at least
ones that are built as part of the project; if no depdency, the .mod file
may not exist before it is used, and if so, scons will abort with an error
("Fatal Error: Cannot open module file 'module1.mod' for reading at (1):
No such file or directory"). With this change, module USE statements are
treated a bit differently. While they are still cached with includes,
the modules are split apart if the cache is found; and the FORTRANMODDIR
when using the file finder to search for files from the scanner.

Rquired fixing test Fortran syntax to use MODULE PROCEDURE in proper
context. Remove use MOD_BAR as you don't use a MODULE PROCEDURE via use
Added new test case: tests that module inclusion generates the proper
dependencies, so that if the alphabetically-first file uses other modules,
thse module files are built first.

Fix to test framework to support file_fixture() taking a list for dstfile.
The docstring says it does, but it didn't do the joining that was done
for srcfile.

Signed-off-by: Mats Wichmann <mats@linux.com>