Maint: remove dead code from sconsign
The Utilities/sconsign.py code used to do its own importing of the
suitable module to read the sconsign file; for a while now it leaves that
to the sconsign module, but some remnants of the old code were still
present, which caused checkers and IDEs to grumble about usage of the
"imp" module, which is removed from Python as of 3.12.  Dropping this
(uncalled) code and associated comments quiets the grumbles.

Also tweaked manpage just a bit.

Signed-off-by: Mats Wichmann <mats@linux.com>