Lock creation of CacheDir config
When creating a new CacheDir, the config file is created in
exclusive mode, but there's a timing window before the json dump
to the file completes when another thread could read the config
because it exists - but get a JSONDecodeError because it hasn't
finished writing yet.  Add locking so the readers will have to
wait until the writer is done.

Fixes #4489

Signed-off-by: Mats Wichmann <mats@linux.com>