Improve env.Dump('json') a bit
When doing a Dump in json mode, custom SCons classes are not examined
by default (unlike the standard pprint() Dump), and end up in the
un-serialized bucket.  Mostly, that's fine, but the BUILDERS dictionary
contains valueable information (which builders actually got configured),
as do the CLVar instances - these contain actual cmdline options that will
be issued.  Update the function called when an object looks unserializable
to recognize subclasses of UserDict (which covers BuilderDict) and
UserList (which covers CLVar) and return their .data field.

Also tell the json encoder to sort the keys.

The adjacent function in the source file got a small tweak (non-functional).

Fixes #4493

Signed-off-by: Mats Wichmann <mats@linux.com>