Give more control to docs build  [skip appveyor]
The documentation build can now selectively disable doing the APIdocs
build (which is slow, and requires Sphinx and some extensions to be
installed); and skip building pdf versions of docs (which requires fop
or equivalent to be installed). Both can be disabled by using a csv:

scons doc SKIP_DOCS=pdf,api

TODO: pdf disabling is only partly working. It does not build the
pdf version of the api docs, but does build pdf manpages and user guide.
The latter are not "installed": there is no build/doc/PDF, but there
are pdfs in the build directory, build/doc/man and build/doc/user.

This is part two of the doc build change (begun in #4492)

Signed-off-by: Mats Wichmann <mats@linux.com>