Fix isabs() check in NodeInfo for Py 3.13
Python 3.13 (alpha) changes the behavior of isabs() on Windows. SCons
has places where it does splitdrive on an absolute path, then checks
if the path part is absolute - this answer is now False, which caused
some interesting test fails.  Do the check on the original path to get
a more accurate answer.  There may be more subtle issues with the Python
change, but first fix the ones we can see.

Simplify the setup of _my_splitdrive() a bit: every caller is supposed to
check do_splitdrive but a couple of locations did not. Remove the special
check for UNC support, all Python versions SCons runs on do UNC handling,
so just eliminate that bit.

Fixes #4502, #4504.

Signed-off-by: Mats Wichmann <mats@linux.com>