Tools update, maninly documentation
Two small code changes:
(1) if Environment() is called with a tools kwarg, reset TOOLS. Since it
wasn't used in the init, any previous contents would be invalid for this env.
(2) When adding the name of a tool to TOOLS, use AppendUnique rather
than Append, so we get a uniqued list.  This impacted one unit test,
which uses a mock Envirionment with no AppendUnique method, so add one -
the result is not used so the fact it's not doing any uniquing is ok.

The rest is doc rewording and reorganization, and a little extra
checking in the site_dir test.

Signed-off-by: Mats Wichmann <mats@linux.com>