Back off test framework exception handling
The try block in TestCmd.start() is removed, it interferes with
other things.  TestCommon.start is supposed to handle exceptions,
so inserting more stuff at a lower level (parent class) made
a mess of the framework tests.

Clean up a complaint with later Pythons (3.12+), and include
(incomplete) new regex for Python 3.13+.

Signed-off-by: Mats Wichmann <mats@linux.com>