Extend pdf-skipping to man and user builds.
An earlier change extended SKIP_DOC as a CLI argument at the top
level to allow granularity - can specify pdf and/or api to skip those.
Skipping the pdf build didn't entirely work, as the build re-invokes
SCons which is then a fresh context. That information is now passed on
the command line to those two builds, and if invokved directly (as Debian
packaging does), SKIP_PDF=1 can be added on the command line.

Signed-off-by: Mats Wichmann <mats@linux.com>