Maintenance: Python version usage
Simplify/clarify some usage of Python version strings.  Dropped a couple
which refer to unsupported versions.

While looking at version usage, it became clear one of the SWIG tests
wasn't doing the right thing - it has a special case for Windows, where
the way the path to the Python library is constructed would only work
for the python.org release, not the Store version.  Since we compute
this information correclty in the test framework, and the test already
called that routine - before ignoring the results - just use that instead.
Though it turns out the framework didn't quite do the right thing either,
returning a bare lib name without the library suffix - fixed that, too.

Signed-off-by: Mats Wichmann <mats@linux.com>