scons-time: add filter to tarfile extract call.
Python 3.13.0b1 added a warning when the extraction filter is
not specified.  The filter aregument (to extract() and extractall())
was added in 3.12, but with no noise. Once the warning was added, the
scons-time tests began to fail - this is a continuation of the work Red
Hat did for RHEL 8 and RHEL 9 starting with Python 3.9.  Supplying the
filter quiets the warning, which also worked for the RHEL case.

The tarfile and zipfile usage now uses a context manager (both
objects can be used this way since Python 3.2 or so).

Signed-off-by: Mats Wichmann <mats@linux.com>