Tweak some things about Command() docs
Adjust the description of the Command function (and builder entry),
mainly to try out new terminology.  We've said it's not really a builder,
but it feels better to me to admit it is *exactly* a builder call, just
not to one of the existing named builders. So the term introduced is
"anonymous builder". Changes in docstring for Command, too, plus a
little tweak to kwarg processing.

This change actually started with reformatting some examples in the
manpage, which used Command, which used the chdir argument which didn't
seem to be mentioned in the description of how to use Command, which
lists four keyword args and says all the rest are treated as construction
variable settings (not true for chdir, which isn't one of the four).
That led to thinking about rewording.

The original code reformatting changes, no longer directly related to the
current change subject, are left in just because they seem low burden,
tough could back them out if it makes reviewing to hard.

A couple of other small tweaks:

* some places referred to a section named "File and Directory Nodes"
  but it's no longer named that.  Actual links to the section updated
  fine when it was renamed, but in Environment.xml, which produces shared
  content, you can't use manpage-specific links, so the _textual_
  references needed updating.
* the code example for a warning was moved inside the warning -
  the effect on output is just an indent to match the warning.
* Configure used both terms "custom checks" and "custom tests", now
  only the former; plus a link reference added.

Signed-off-by: Mats Wichmann <mats@linux.com>