Fix some more regex strings
Building the userguide examples with Python 3.12 led to SyntaxWarning
lines appearing in the output. The reason was a regex string that
contained an escape sequence that is not one of the Python escapes,
and the string wasn't entered in raw-string format - 3.12 compains
about this. This sort of problem has been fixed elsewhere, but missed
some cases in the tools. Continuing on with the rule "if it's a regex,
use raw string format".

Signed-off-by: Mats Wichmann <mats@linux.com>