Update to current fastest env setitem
Updated and re-ran benchmark tests for SubstitutionEnvironment.__setitem__.
Added test for new (Python 3.0) string.isidentifier() method.
That's actually exactly what we want, and it times out as the fastest
method when combined with a membership test if the variable is
already defined.

Tweaked some comments about this performance consideration,
and did other updates in bench/.

Signed-off-by: Mats Wichmann <mats@linux.com>