Enhancement: Dump() takes multiple args now
env.Dump previously either printed the whole dict of construction
variables (with no args), or a single value (with one variable name
argument).  Now it takes a varargs specifier so you can give any number
of consvar names.  All returned strings are now in dict form, including
the single-arg case which previously just returned the value matching
the key, not a dict with a key:value pair. This is a slight ABI change,
but should not affect any actual scripts since the output is intended
for human consumption, not for programmatic use - env.Dictionary()
can be used to fetch construction vars for programmatic use (in fact,
Dump is a consumer of Dictionary's output).

Signed-off-by: Mats Wichmann <mats@linux.com>