Restore usage of "from SCons.Variables import *`
This (unadvertised) import turns out to have real-world usage,
and there's no real reason to prohibit it, although we do suggest
using `from SCons.Script import *` as a more complete version.
SCons 4.8.0 introduced an `__all__` in the variables module;
add the five variables types to this so the listed import will
work as expected.

Signed-off-by: Mats Wichmann <mats@linux.com>