Switch scons build to use setuptools version hook  [skip appveyor]
The stanza we used, still published in the Python packaging docs,
is now considered the least desirable of the alternatives for keeping
a "single package version", so update to a better one.

Note there's what's considered a "better" one than this, but it requires
Python 3.8 ("provisional" until 3.10) or a backport of importlib.metadata.

Signed-off-by: Mats Wichmann <mats@linux.com>