Fix bug with unique adds and delete_existing
AppendUnique and PrependUnique, when called with delete_existing true, had
a small logic flaw: it might remove an existing value if the value to be
added is a scalar (string, most likely) and there was a substring match.
The code needs to do an equality test, not a membership test.

Unit tests updated and got some reformatting, plus dropped a duplicate
definition of reserved_variables.

Signed-off-by: Mats Wichmann <mats@linux.com>