Modify the Builder Methods manpage section
In clarifying the detail on when a builder's target can be deduced,
it seemed like the concept of a single-source builder ought to be
introduced in this context. So now those builders we know are
single-source have that notation in their entry in this section.

Also mentions that pseudo-Builders may have a diffeent calling sequence.

A chunk of text about dependencies previously appeared after the
listing of builder methods, where it would be hard to spot. Moved up
into the text before the listing, and integrate in a bit more smoothly.

A couple of very minor typing things adjusted along the way - mainly
using single_source consistently as a bool.

Signed-off-by: Mats Wichmann <mats@linux.com>