Tweak SConsOptions docstrings
Indicate more clearly (for the API docs) that something is an SCons
addition/modification, since that doc shows members inherited from the
optparse class - this was always easier to tell in the code than in the doc.

A few other minor rearrangements and comments and a few type annotations.

Does not change any behavior.

Signed-off-by: Mats Wichmann <mats@linux.com>