PackageVariable now returns the default on "true"
In all doc versions until 4.8.0, PackageVariable had wording like:
"The option will support the values yes, true, on, enable or search,
in which case the specified default will be used", but the code didn't
actually do that, it just returned True. With this change it now returns
the default value, with a slight tweak - if the default is one of the
spelled out enabling strigs, it returns the boolean True instead.

The indication that the default is produced if a truthy string is given
is restored to the manpage (it was never dropped from the User Guide).

Signed-off-by: Mats Wichmann <mats@linux.com>