Tweak Clean and NoClean manpage entries  [skip appveyor]
Motivated by an issue that pointed out the argument names differ between
doc and implementation, adjusted the wording a bit.  Clean() used to
suggest you could use a whole bunch of args in one call ("Multiple files
or directories should be specified either as separate arguments to the
Clean method, or as a list") but as it's written to take exactly two
positionial arguments that wasn't correct - only the list approach
works.

Fixes #4638

Signed-off-by: Mats Wichmann <mats@linux.com>