Turn SCons.Variables.Variable into a dataclass
Also correct one long-standing minor inconsistency: if a list of names
for a variable is given, it is split into name and aliases, with the
name not duplicated into the aliases. If only one name is given, it is
both the name and the only entry in the aliases. All of the usages of
aliases combine them back together anyway, like:

    if arg in option.aliases + [option.key,]:

So there's no need for the behavior of the one-name form including the
name in aliases. One test did need to change for this - the test of a
custom help formatter function was wired to accept the old form, but
only because of the way the custom formatter in the test was written,
so this is a self-contained problem.

Signed-off-by: Mats Wichmann <mats@linux.com>