Further tweak CPPDEFINES repleacement in scanner
Replacement is now limited to five passes, to avoid going into an endless
loop in pathlogical cases where there are three or more macros that
circularly refer to each other. No error is reported in this case.

Replacement is now only done for the otional C Conditional scanner,
not for the classical C scanner.

Signed-off-by: Mats Wichmann <mats@linux.com>