Update Pre/PostAction Alias note  [skip appveyor]
The initial version of PR #4872 proposed to document that pre- and
post-actions on an alias have no effect if the alias has no action
itself. Subsequently, PR #4874 was merged to change this behavior so
that statement is backed out, but the other proposed doc improvements
in this area are retained.

Signed-off-by: Mats Wichmann <mats@linux.com>