Fix running individual test files when ninja is not installed
The code in testing/framework/TestSCons.py attempted to handle an
ImportError if ninja is not available.  However, when running individual
test files from the scons/test/ directory, this directory is included as
the first entry in sys.path.  When this happens, the `import ninja`
statement succeeds, finding the scons/test/ninja/ directory and treating
it as a package.  This results in an AttributeError being thrown later
attempting to access `ninja.BIN_DIR`, rather than an ImportError.

I have confirmed that this change now allows `./runtest.py test/Help.py`
to succeed, even when ninja is not installed.