Fortran dialects default FLAGS and PATH if not set
The manpage has always said that for a given dialect (e.g. F90), to set
FORTRANFLAGS and FORTRANPATH unless you need values that are unique
to just that dialect, in which case you can set the dialect-specific
ones (e.g. F90FLAGS and/or F90PATH). However, the implementation never
actually picked up the generic settings if the dialect-specific settings
were not given. This change now picks up the defaults if the dialect
does not have a value set.

Signed-off-by: Mats Wichmann <mats@linux.com>