Fortran dialects default FLAGS and PATH if not set
The manpage has always said that for a given dialect (e.g. F90), to set
FORTRANFLAGS and FORTRANPATH unless you need values that are unique
to just that dialect, in which case you can set the dialect-specific
ones (e.g. F90FLAGS and/or F90PATH). However, the implementation never
actually picked up the generic settings if the dialect-specific settings
were not given. This change now picks up the defaults if the dialect
does not have a value set.

After initial change to default to generic dialect's flags/paths, if a
test did an Append to *FLAGS, it would do string concatenation with
the un-substituted default, as that looks like a string.  That is,
if $SHF90FLAGS is '$SHFORTRANFLAGS', then appending '-x' became
'$SHFORTRANFLAGS-x' rather than ['$SHFORTRANFLAGS', '-x'].  This failed
tests on Windows, though for some reason I'm not clear on, not on Linux.
Turning the initial assignment to a CLVar takes care of this.

Signed-off-by: Mats Wichmann <mats@linux.com>