Fix Fortran dialect FLAGS default
After the change to default to generic dialect's flags/paths, if a
test did an Append to *FLAGS, it would do string concatenation with
the un-substituted default, as that looks like a string.  That is,
if $SHF90FLAGS is '$SHFORTRANFLAGS', then appending '-x' became
'$SHFORTRANFLAGS-x' rather than ['$SHFORTRANFLAGS', '-x'].  This failed
tests on Windows, though for some reason I'm not clear on, not on Linux.
Turning the initial assignment to a CLVar takes care of this.

Signed-off-by: Mats Wichmann <mats@linux.com>