Interactive mode batch + src/tgt clearing
On Windows, where batch mode exists, the changed_{sources,targets} lists
become important. In interactive mode, these were not cleared properly,
leading to incorrect results. Moved them into the memo dict, which is
dropped when necessary, to match the other source and target entries.

The testcase is taken from the issue, but needed to apply some of the
tricks for interactive-mode e2e tests, where testing is not entirely
straightforward.

Fixes #3029

Signed-off-by: Mats Wichmann <mats@linux.com>