Clean up CC and CXX FLAGS tests
More consistency, formatting, add test/CXX/SHCCFLAGS.py for symmetry
(test/CC/ had one), rename tests to -live since they use a real compiler,
and also add that note to docstring. Remove some filename suffix
fiddling that wasn't needed because in this usage, SCons figures it out -
and we wnat that as part of the test.

Signed-off-by: Mats Wichmann <mats@linux.com>