Fix CacheDir initialization so works on 3.7 also
Although there is no indication of a change in this area of the tempfile
module, it seems if the temporary directory has been renamed so the
original no longer exists, the context manager fails on 3.7, though
not on any later Python versions (maybe a bugfix rather than a planned
behavior change?). Now use mkdtemp instead, and clean it up by hand
if the rename failed (also use os.replace instead of os.rename,
so it will actually fail if the directory existed and is nonempty,
on Linux os.rename doesn't fail in this case).

Fixes #4694

Signed-off-by: Mats Wichmann <mats@linux.com>