Exp: try to make the Windows Docbook tests better
Try to fix Windows fails on Docbook tests in case xsltproc is found.
On both GitHub Actions and AppVeyor, it's found as part of StrawberryPerl,
which is part of the default install. Intermittent fails seem caused by
network issues (whether actual problems, or forced revectoring to https
not supported by xsltproc according to some internet issue reports), so
propagate two xsltproc flags for this that were in three of the "live"
(named "cmd" here) tests, but not the others. Also made other parts of
the setup of these tests more consistent (passing found xsltproc
wasn't done)

Signed-off-by: Mats Wichmann <mats@linux.com>