Fix tests for use with Python 3.14
As with every Python release, a few tests have to be fiddled to
work with that release. SCons/ActionTests is always one of those,
as it has dictionaries mapping (major, minor) version to corresponding
bytecode strings. The only other outstanding one this go-round was
an expection of the exception messages for a divide-by-zero,
which has changed slightly again.

Signed-off-by: Mats Wichmann <mats@linux.com>