Add manpage .1 files to doc tarball  [skip appveyor]
The three generated roff-style (.1) manpages are included in the `scons-doc` tarball which is generated at release time and uploaded to `scons.org/doc`. This provides a more natural place to grab them from if someone wants to download this tarball (its existence is not currently advertised, though one distro packager is known to use id; normally it is used to populate the documenations matrix for each version.

No scons changes involved, and not user-visible, since you have to grab
the tarball explicitly.

Signed-off-by: Mats Wichmann <mats@linux.com>