MSVC: Import POWERSHELL_TELEMETRY_OPTOUT, PSDisableModuleAnalysisCacheCleanup, and PSModuleAnalysisCachePath into the limited msvc environment when defined.
Environment variables added to msvc environment import list:
* POWERSHELL_TELEMETRY_OPTOUT: opt-out of telemetry.
* PSDisableModuleAnalysisCacheCleanup: disable checking for modules that no longer exist when writing module analysis cache.
* PSModuleAnalysisCachePath: path and filename used to cache data about modules and their cmdlets.

The PSModuleAnalysisCachePath is defined in windows runners 2019, 2022, and 2025.  The significant delays experienced in the 2022 and 2025 runners when running the msvc batch files due to vcpkg with powershell 5 and without PSModulePath defined appear to be eliminated when the powershell module analysis cache path is propagated to the msvc environment.

Additional changes:
* Move the PS7 path after the PS5 path.
* Add the elapsed execution time of the msvc batch files to the debug log.