Add test showing missed dependencies
The LaTeX scanner is not picking up input files from a subimoprt if they
are not in the subimport's directory but are on TEXINPUTS contrary to
how LaTeX will search.  It additionally filters out files that have the
same name even if they are in separate directories.  This demonstrates
the behavior.