Fix ninja problems with quotes and ListActions
Ninja tool now quotes targets (if necessary) when calling back to
SCons - both in the POST request constructed to contact the
daemon, and in the command eventually issued from the deamon.
Initial suggestion from Julien Pommier.

Fixes #4730

Ninja tool is also adjusted to recognize and emit a CMD tagged rule
in the case of a list action (... action=[Something, LaterSomething]).
CMD refers to functions that have special handlers that need the
source and target as arguments, in this case the Copy function
(at least) was being categorized as a GENERATED_CMD, which gets
no arguments, and so the copy failed.

Fixes #4731

Signed-off-by: Mats Wichmann <mats@linux.com>