FS unit test fixes for UNC paths
Managed to fail a number of UNC path tests by moving everything,
even the %TEMP% directory to a dev drive from the C drive.
A structuring problem meant some paths weren't trimmed of
trailing slashes as they were intended to - we just do a
string comparison in the cases, not a path-aware comparison,
so this was enough for an error.  Resynced three similar tests,
added some docstrings and comments, etc.

Signed-off-by: Mats Wichmann <mats@linux.com>