Fix a few niggles releaed to "building" scons  [skip appveyor]
Typo in SConsDoc.py caused Sphinx warn.  Sphinx was also issuing this notice:

  Converting `source_suffix = '.rst'` to `source_suffix = {'.rst': 'restructuredtext'}`.

so updated the conf file to reflect that evolution.

Some tweaks to pyproject.toml to reflect some evolution as well.
Build was issuing this:

  SetuptoolsDeprecationWarning: `project.license` as a TOML table is deprecated
!!
  Please use a simple string containing a SPDX expression for `project.license`.
  You can also use `project.license-files`. (Both options available on setuptools>=77.0.0).

Plus, License in Classifiers is deprecated, as is license-files in
the [tool.setuptools] section (which moves to the package section.

Improved, I hope, the package-finder specification.

Added optional dependencies section to match the external
requirement-dev.txt and requirements-pkg.txt files. This is supposed
to allow installing "in one go", like (currently untested):

  python -m pip install scons[dev]

Signed-off-by: Mats Wichmann <mats@linux.com>