Update Configure docs, again  [skip appveyor]
Tweak the wordings of CheckLib, CheckLibWithHeader, CheckType,
CheckTypeSize.

Add id markers to all the configure methods so they can be linked to.

Also added a docstring to the Node "Attrs" inner class - this relates to
Configure because it adds attributes to Nodes directly, but Attrs
uses __slots__ - seemed worth a brief clarification.

Signed-off-by: Mats Wichmann <mats@linux.com>