Missing builders produce more targeted message
Simply traps the AttributeError by adding a __getattr__ to
SCons.Environment.Base and prints a message that gives some ideas.
Behavior does not change, and there's no cost - we don't reach this code
unless we're already handling an exception for a missing attribute.

An initial thought was to use methods named in
SCons.Script.GlobalDefaultBuilders to trigger the more detailed message,
as a warning (off by default, but can be enabled).  After letting it sink
in for a while, that didn't seem to be any real benefit to that extra
step, nor for hiding the message in a warning. So now consider this as a
"better error message" and nothing more.

Another intial thought was to reclassify this as a UserError - you
have either misspelled a method, or it's not in the execution
environment path, or the tool never ran to set it up, all of which
are the developer's problem. But enough code depends on this being
an AttributeError specifically that several things broke - would have
been a lot more work than a "quick fix".

Doesn't do anything for the case of missing "global function" type
builders, they have their own processing logic which raises a NameError.

This is in response to multiple issue reports on confusing errors (#863
and #1671 are open, #1437, #1442, #1873, #1895, #1007 closed as dupes -
at least), but does not really "fix" them, that appears to require some
rearchitecting of how the whole tools+builders+environment-methods
scheme works to gather early information so you can really tell what is
unavailable that was expected to be.

Three tests for removed features that expected a specific message on
attempt to access that feature were reworked as a result.

Signed-off-by: Mats Wichmann <mats@linux.com>