Add manpage entry for --cache-dir + tweaks
Tweak: intermediate directories are now made so the temporary
cachedir creation is less likely to fail.

Tweak: if debug output is selected and sent to a file, the
file is now opened in append mode. This fixes a problem of lost output
if multiple cachedirs are used in one run (presumably rare except
for testing scenarios), but isn't an ideal solution for the issue.

Signed-off-by: Mats Wichmann <mats@linux.com>