Improve Literal class and a few cleanups
It wasn't possible to do in/not in checks on strings wrapped in a
Literal(), because it had no __contains__ method. Added.

SCons.Util included MappingView in SequenceTypes, but a MappingView is not
iterable so is not correct here. Instead added the three dict *contents*
views (Keys, Values, Items), which are iterable.

The rest is fiddly cleanup: minor typing, docstring and import
changes; do the standard "test cleanup" on test/Subst/Literal.py.

Signed-off-by: Mats Wichmann <mats@linux.com>