Use "in" and "not in" instead of find()
A maintenance edit: most instances of str.find() used for membership
checks like "if foo.find('bar') != -1" are replaced with "in" or "not in",
or in a few cases, with the "startswith()" method.  The vendored docbook
utilities were not changed in this round though they had a couple of these.

Signed-off-by: Mats Wichmann <mats@linux.com>