Add preliminary test showing failure
This test (should) generate two identical LaTeX input files: main.tex
and main.extra.tex.  The should compile to the same content, but SCons
strips accidentally strips the '.extra' from the second preventing it
from finding the bibliography file.  The test needs to be checked on a
non-Windows system because Miktek does not support the ~ in the
shortened path name to the temp directory.