Try to call Override less
Eliminate a kwarg 'explain' from the call to Builder() for Mkdir.  Arg is
not used (maybe it was in the past, though have not found the evidence
for this), which meant we had an override dict in this case. Suprisingly,
this ended up in the override logic a *lot* in a build of a large sample
project - over 2000 extra calls to scons_subst_once that weren't needed.

Builder and Executor both have places where they call Override().
While the Override factory indeed returns quickly if the override dict
is empty, avoid the function call entirely in this case by checking
(there was an old TODO comment to this effect in Builder.__init__.py).

The Executor init method declared multiple paramters with mutable
defaults, which means they "persist". While there's no evidence this
is calling a real problem, eliminate part of this footgun by defaulting
overridelist and builder_kw to None and checking for that (targets and
sources still default to a mutable empty list, a TBD perhaps).

Signed-off-by: Mats Wichmann <mats@linux.com>