Add some basic unit tests for file locking
SCons/Util/filelock.py was untested, in anticipation of making use
of it for additional cache-related items, introduced a few basic unit
tests. More could be done but it's better than nothing.

The locking module changed only minimally - separate exceptions are now
used for two different kinds of lock failure - this is mainly for the
tests, a user can continue to just look at the SConsLockFailure base
exception. The lock class repr is alse beefed up a bit.

Signed-off-by: Mats Wichmann <mats@linux.com>