Fix Sphinx build, again  [skip appveyor]
The release of Sphinx 9 brought a rewritten autodoc plugin, which now
fails on one use of the SCons.Util.Null class that it's been issuing a
warning for since we stated using Sphinx (could never find the cause).
Found the location - the default value of the ProgressObject is Null,
and that caused the problem. Not entirely sure why, but that usage doesn't
need a Null, since it checks for a sentinel, so just switch it to use None.

Also comment out one unused feature in the Sphinx config that issued
a warning.

Signed-off-by: Mats Wichmann <mats@linux.com>