Improve covarage of API doc build  [skip appveyor]
Tell Sphinx to Ignore setting of __all__ in SCons.Variables (would ignor
elsewhere but that's the only current user), it was skipping some of
the content of Variables previously.  Don't show inherited members for
SCons.Script.SConsOptions, there's little docstring content for stdlib
optparse, and this is a place where it's clearer if we see only the
things SCons extends.  Tweak a few docstrings in SConsOptions, too.

Signed-off-by: Mats Wichmann <mats@linux.com>